add error check for normalized vector
