+++ include/parse.h
@@ -6,7 +6,7 @@
 /*   By: naharagu <naharagu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/25 22:04:47 by naharagu          #+#    #+#             */
-/*   Updated: 2023/05/29 13:20:51 by naharagu         ###   ########.fr       */
+/*   Updated: 2023/05/29 22:52:15 by naharagu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,6 +59,7 @@ void					parse_light(char **str_array, t_scene *scene);
 //parse_utils.c
 void					free_split(char **split);
 size_t					count_array(char **str_array);
+void					check_normalized(t_vec3 vec);
 t_color					parse_color(char *str);
 t_vec3					parse_vec3(char *str);
